verify if kind exist

+++ kind-cluster/install.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
-curl -Lo ./kind https://github.com/kubernetes-sigs/kind/releases/download/v0.8.1/kind-linux-amd64
-chmod +x ./kind
-sudo mv ./kind /usr/local/bin/kind
+
+if [[ ! -f "/usr/local/bin/kind" ]]
+then
+    curl -Lo ./kind https://github.com/kubernetes-sigs/kind/releases/download/v0.8.1/kind-linux-amd64
+    chmod +x ./kind
+    sudo mv ./kind /usr/local/bin/kind
+fi
+
+
 
 kind create cluster --config kind-config.yaml
-
 kubectl apply -f https://raw.githubusercontent.com/kubernetes/ingress-nginx/master/deploy/static/provider/kind/deploy.yaml
 kubectl wait --namespace ingress-nginx --for=condition=ready pod --selector=app.kubernetes.io/component=controller --timeout=90s
 kubectl create -f metric-server.yaml
